Separated out ConnectedComponents and Coreness traits
