+++ src/dachshund/cnm_communities.rs
@@ -7,17 +7,10 @@
 extern crate nalgebra as na;
 use crate::dachshund::graph_base::GraphBase;
 use crate::dachshund::id_types::NodeId;
-use crate::dachshund::node::Node;
-use na::{DMatrix, DVector};
+use crate::dachshund::node::{Node, NodeBase, NodeEdgeBase};
 use ordered_float::OrderedFloat;
-use rand::distributions::WeightedIndex;
-use rand::prelude::*;
-use rand::seq::SliceRandom;
-use rand::Rng;
 use std::cmp::Ordering;
-use std::collections::{BTreeSet, BinaryHeap, HashMap, HashSet, VecDeque};
-use std::collections::hash_map::Keys;
-use std::iter::FromIterator;
+use std::collections::{BinaryHeap, HashMap, HashSet};
 type Community = HashSet<NodeId>;
 
 #[derive(Clone, Copy, Eq)]
@@ -65,7 +58,7 @@ impl PartialOrd for CNMCommunityMergeInstruction {
 }
 type CNMCommunityMergeInstructionHeap = BinaryHeap<CNMCommunityMergeInstruction>;
 
-pub trait CNMCommunities : GraphBase {
+pub trait CNMCommunities : GraphBase<NodeType = Node> {
     fn get_max_maxheap(
         &self,
         delta_q_maxheap: &HashMap<usize, CNMCommunityMergeInstructionHeap>,
@@ -124,9 +117,10 @@ pub trait CNMCommunities : GraphBase {
         let q0: f64 = 1.0 / (num_edges as f64);
         for (_i, community) in communities.iter() {
             for id in community {
-                for neighbor_id in &self.get_node(*id).neighbors {
+                for e in self.get_node(*id).get_edges() {
+                    let neighbor_id = e.get_neighbor_id();
                     let i: &usize = reverse_id_map.get(&id).unwrap();
-                    let j: &usize = reverse_id_map.get(&neighbor_id.0).unwrap();
+                    let j: &usize = reverse_id_map.get(&neighbor_id).unwrap();
                     let k_i: usize = degree_map[i];
                     let k_j: usize = degree_map[j];
                     let delta_qij: f64 =
