+++ src/dachshund/graph_base.rs
@@ -5,25 +5,26 @@
  * LICENSE file in the root directory of this source tree.
  */
 use crate::dachshund::id_types::NodeId;
-use crate::dachshund::node::Node;
+use crate::dachshund::node::NodeBase;
 use std::collections::HashMap;
-use std::collections::hash_map::Keys;
+use std::collections::hash_map::{Keys, Values};
 
 /// General-purpose trait which indicates the minimum amount of shared context
 /// required between all graph objects. Currently built to accommodate a graph
 /// with "core" and "non-core" ids. A GraphBase is built by a GraphBuilder.
 pub trait GraphBase
 where
-    Self: Sized,
+    Self: Sized, Self::NodeType: NodeBase
 {
     type NodeType;
 
     fn get_core_ids(&self) -> &Vec<NodeId>;
     fn get_non_core_ids(&self) -> Option<&Vec<NodeId>>;
     fn get_ids_iter(&self) -> Keys<NodeId, Self::NodeType>;
-    fn get_mut_nodes(&mut self) -> &mut HashMap<NodeId, Node>;
+    fn get_nodes_iter(&self) -> Values<NodeId, Self::NodeType>;
+    fn get_mut_nodes(&mut self) -> &mut HashMap<NodeId, Self::NodeType>;
     fn has_node(&self, node_id: NodeId) -> bool;
-    fn get_node(&self, node_id: NodeId) -> &Node;
+    fn get_node(&self, node_id: NodeId) -> &Self::NodeType;
     fn count_edges(&self) -> usize;
     fn count_nodes(&self) -> usize;
 }
