+++ src/dachshund/graph_builder.rs
@@ -18,6 +18,7 @@ pub trait GraphBuilder<TGraph: GraphBase>
 where
     Self: Sized,
     TGraph: Sized,
+    TGraph: GraphBase<NodeType=Node>
 {
     fn _new(
         nodes: HashMap<NodeId, Node>,
