+++ src/dachshund/mod.rs
@@ -7,6 +7,8 @@
 pub mod beam;
 pub mod candidate;
 pub mod cnm_communities;
+pub mod connected_components;
+pub mod coreness;
 pub mod error;
 pub mod graph_base;
 pub mod graph_builder;
