+++ tests/simple_graph.rs
@@ -7,6 +7,8 @@
 extern crate lib_dachshund;
 use crate::lib_dachshund::TransformerBase;
 use lib_dachshund::dachshund::cnm_communities::CNMCommunities;
+use lib_dachshund::dachshund::connected_components::ConnectedComponents;
+use lib_dachshund::dachshund::coreness::Coreness;
 use lib_dachshund::dachshund::id_types::NodeId;
 use lib_dachshund::dachshund::input::Input;
 use lib_dachshund::dachshund::output::Output;
