+++ tests/tests.rs
@@ -7,7 +7,6 @@
 extern crate lib_dachshund;
 
 use lib_dachshund::dachshund::candidate::Candidate;
-use lib_dachshund::dachshund::cnm_communities::CNMCommunities;
 use lib_dachshund::dachshund::error::{CLQError, CLQResult};
 use lib_dachshund::dachshund::id_types::{GraphId, NodeId};
 use lib_dachshund::dachshund::line_processor::LineProcessorBase;
